Refactor tests as well
